+++ packages/graphql-hooks/src/middlewares/apqMiddleware.ts
@@ -1,6 +1,6 @@
 import { Sha256 } from '@aws-crypto/sha256-browser'
 import { Buffer } from 'buffer'
-import { MiddlewareFunction } from '../types/common-types'
+import { APIError, MiddlewareFunction } from '../types/common-types'
 
 export async function sha256(query) {
   const hash = new Sha256()
@@ -18,6 +18,25 @@ type APQExtension = {
   }
 }
 
+function isPersistedQueryNotFound(error: APIError) {
+  if ((error?.fetchError as any)?.type === 'PERSISTED_QUERY_NOT_FOUND') {
+    return true
+  }
+
+  let errors = error?.graphQLErrors ?? []
+
+  if (error.httpError) {
+    try {
+      const body = JSON.parse(error.httpError.body)
+      errors = errors.concat(body.errors ?? [])
+    } catch {
+      return false
+    }
+  }
+
+  return errors.some(e => e.message === 'PersistedQueryNotFound')
+}
+
 /**
  * AutomaticPersistedQueryMiddleware - must be last in the middleware list
  * @param {function} makeRequest
@@ -27,7 +46,6 @@ export const APQMiddleware: MiddlewareFunction<APQExtension> = async (
   { operation, client, resolve, reject },
   next
 ) => {
-  const ERROR_PERSISTED_QUERY_NOT_FOUND = 'PERSISTED_QUERY_NOT_FOUND'
   try {
     operation.extensions = {
       ...operation.extensions,
@@ -52,19 +70,8 @@ export const APQMiddleware: MiddlewareFunction<APQExtension> = async (
 
     const { error } = res
 
-    if (!error.fetchError && !error.graphQLErrors) {
-      throw error
-    }
-
-    const persistedQueryNotFound = error.fetchError
-      ? (error.fetchError as any).type === ERROR_PERSISTED_QUERY_NOT_FOUND
-      : error.graphQLErrors?.some(
-          gqError =>
-            gqError?.extensions?.code === ERROR_PERSISTED_QUERY_NOT_FOUND
-        )
-
     // If a server has not recognized the hash, send both query and hash
-    if (persistedQueryNotFound) {
+    if (isPersistedQueryNotFound(error)) {
       next()
     } else {
       throw error