+++ packages/graphql-hooks/test-jsdom/unit/GraphQLClient.test.ts
@@ -1,16 +1,12 @@
 import { GraphQLClient } from '../../src'
 import { TextEncoder } from 'util'
 import fetchMock from 'jest-fetch-mock'
-import { FormData, File as FormDataFile } from 'formdata-node'
-import { Readable } from 'stream'
+import { FormData, File } from 'formdata-node'
 import {
   createMockCache,
   createMockResponse,
   createMockSubscriptionClient
 } from '../utils'
-
-// workaround for https://github.com/octet-stream/form-data/issues/50
-const { fileFromPathSync } = require('formdata-node/lib/cjs/fileFromPath')
 
 global.TextEncoder = TextEncoder
 
@@ -462,51 +458,6 @@ describe('GraphQLClient', () => {
 
       it('does not set Content-Type header', () => {
         expect(fetchOptions.headers).not.toHaveProperty('Content-Type')
-      })
-    })
-
-    describe('with files in Node JS', () => {
-      const originalFormData = global.FormData
-
-      const client = new GraphQLClient({ ...validConfig, FormData })
-      const file = fileFromPathSync('test/mocks/sample.txt')
-
-      const operation = { query: '', variables: { a: file } }
-      const fetchOptions = client.getFetchOptions(operation)
-
-      beforeAll(() => {
-        //@ts-ignore
-        delete global.FormData
-      })
-
-      afterAll(() => {
-        global.FormData = originalFormData
-      })
-
-      it('sets body conforming to the graphql multipart request spec', () => {
-        const actual = fetchOptions.body as FormData
-
-        expect(actual).toBeInstanceOf(FormData)
-        expect(actual.get('operations')).toBe(
-          '{"query":"","variables":{"a":null}}'
-        )
-        expect(actual.get('map')).toBe('{"1":["variables.a"]}')
-        const actualFile = actual.get('1') as unknown as File
-        expect(actualFile).toBeInstanceOf(FormDataFile)
-        expect(actualFile.name).toBe('sample.txt')
-      })
-
-      it('does not set Content-Type header', () => {
-        expect(fetchOptions.headers).not.toHaveProperty('Content-Type')
-      })
-
-      it('throws if no FormData polyfill provided', () => {
-        const client = new GraphQLClient(validConfig)
-        const operation = { query: '', variables: { a: new Readable() } }
-
-        expect(() => client.getFetchOptions(operation)).toThrow(
-          'GraphQLClient: FormData must be polyfilled or passed in new GraphQLClient({ FormData })'
-        )
       })
     })
   })